Move GetOutpost2Directory function
Use GetOutpost2Directory to prevent NetFixClient logger from logging in environment path when a debugger is attached. (This likely won't be a problem in the future if we remove the local logger)

Use OP2Internal specific function when getting the default protocol index. Simpler call and still works proper with debugger attached.
